api/pyproject.toml
[project]
name = "bivouac-api"
version = "0.1.0"
description = "The basecamp for adventure, nature & extreme sports documentaries - API"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Bivouac.tv Team" }]

dependencies = [
    # Django core
    "django>=5.1,<6.0",
    "djangorestframework>=3.15,<4.0",
    "django-cors-headers>=4.6,<5.0",
    "django-filter>=24.3,<25.0",

    # Authentication
    "dj-rest-auth[with_social]>=7.0,<8.0",
    "djangorestframework-simplejwt>=5.3,<6.0",

    # Database
    "psycopg[binary]>=3.2,<4.0",
    "dj-database-url>=2.3,<3.0",

    # Storage & Media
    "django-storages[s3]>=1.14,<2.0",
    "pillow>=11.0,<12.0",

    # Configuration
    "django-environ>=0.11,<1.0",

    # Utils
    "django-extensions>=3.2,<4.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3,<9.0",
    "pytest-django>=4.9,<5.0",
    "pytest-cov>=6.0,<7.0",
    "factory-boy>=3.3,<4.0",
    "faker>=33.0,<34.0",

    # Code quality
    "ruff>=0.8,<1.0",
    "pre-commit>=4.0,<5.0",
    "mypy>=1.13,<2.0",
    "django-stubs>=5.1,<6.0",
    "djangorestframework-stubs>=3.15,<4.0",

    # Debug
    "django-debug-toolbar>=4.4,<5.0",
    "ipython>=8.30,<9.0",
]

prod = [
    "gunicorn>=23.0,<24.0",
    "whitenoise>=6.8,<7.0",
    "sentry-sdk[django]>=2.19,<3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["apps", "config"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
    "S",      # flake8-bandit (security)
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # assert (we use it in tests)
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
section-order = ["future", "standard-library", "django", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
django = ["django", "rest_framework"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--strict-markers",
    "--reuse-db",
    "-ra",
]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.coverage.run]
source = ["apps"]
branch = true
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
